Creates KeyValueElement for fields and subsequent querystring types.
