+++ pactum/base.py
@@ -22,3 +22,19 @@ class Element:
         child_attr = getattr(self, self._children_name, [])
         child_attr.append(child)
         child.parent = self
+
+
+class KeyValueElement(Element):
+    def __init__(self, name=None, type=None, required=None, **kwargs):
+        super().__init__(**kwargs)
+        if required is None:
+            required = getattr(self, 'required', False)
+        self.required = required
+
+        if name is None:
+            name = getattr(self, 'name', '')
+        self.name = name
+
+        if type is None:
+            type = getattr(self, 'type', self.__class__)
+        self.type = type
